requirements.txt
Flask-mysqldb
venv